refactor: splitting omni-connector into multiple smaller connectors (#48)
* refactor: scratch implementation of refactored version of sdk

* refactor: made solana-connector rely on near-connector

* chore: removed useless lib.rs files

Since we can directly set path for a lib file, then there's no need to create a lib file just to import a module

* fix: manually synced with main branch

* fix: resolved issues after merge

* chore: scratch implementation of bridge-clients

* Implement a few near<->evm operations (#52)

* feat: added wormhole support

Made an abstraction for init&fin transfer as well

* chore: bumped spl crates version

* feat: support for multiple evm chains

* chore: added abstract method for calling `sign_transfer` on NEAR chain

* chore: added abstract method for calling `claim_fee` on NEAR chain

* chore: added abstract method for calling `sign_claim_fee` on NEAR chain

* feat: synced with latest omni-types

* feat: added `get_required_storage_deposit` method on NEAR chain

* fix: static analysis

* feat: added `get_token_id` method on NEAR chain

* feat: support multiple OmniAddress variants

* feat: added `get_native_token_id` method on NEAR chain

* feat: finished abstraction (#54)

* feat: finished abstraction

* feat: included solana into omni abstraction

* feat: made solana's fields optional

This can be useful if user won't use full potential of bridge-client, but only some methods which don't require all fields like wormhole or keypair

* chore: provided a bit more consistency between different bridge clients

* chore: addressed some clippy warnings

* chore: updated some crates

* chore: decreased attached gas

* chore: moved attached gas/deposit to consts

---------

+++ bridge-sdk/near-rpc-client/src/near_rpc_client.rs
@@ -87,7 +87,7 @@ pub async fn get_last_near_block_height(server_addr: &str) -> Result<u64, NearRp
     };
 
     let block_info = client.call(request).await?;
-    Ok(block_info.header.height as u64)
+    Ok(block_info.header.height)
 }
 
 pub async fn get_block(
